requirements.txt
fastapi>=0.117.0,<0.118.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
rembg[cpu]>=2.0.67
pillow>=10.1.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.2
